requirements.txt
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
beautifulsoup4==4.13.4
blinker==1.9.0
brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
crispy-bootstrap4==2025.6
cryptography==45.0.6
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
django==5.2.5
django-crispy-forms==2.4
et-xmlfile==2.0.0
fastapi==0.116.1
flask==3.1.2
fonttools==4.59.1
fpdf==1.7.2
fpdf2==2.8.4
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.1
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.5
narwhals==2.1.2
numpy==2.3.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
playwright==1.54.0
plotly==6.3.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyhanko==0.29.1
pyhanko-certvalidator==0.27.0
pyparsing==3.2.3
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.2
reportlab==4.4.3
requests==2.32.5
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.47.2
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
